Apply prettier diff to ng editor as well
Now the server-side diff is not used at all, so it could be removed.
Might want to be removed a bit of time after this change gets merged
in the case of caching the old ng editor that still depends on the
server-side diff.
