+++ css/hole-ng.css
@@ -1,3 +1,10 @@
+/**
+ * NOTE: These includes are hardcoded in render.go, not inlined automatically
+ * from comments like the *.js `include`s
+ */
+ 
+/* include hole-diff.css */
+
 .term-fg31 { color: #c00 }
 .term-fg32 { color: #490 }
 .term-fg33 { color: #ca0 }
