+++ js/hole.js
@@ -1,6 +1,6 @@
 import CodeMirror from './_codemirror-legacy';
 import strlen     from './_strlen';
-import * as Diff  from "diff";
+import diffHTML   from './_diff';
 
 const chars              = document.querySelector('#chars');
 const darkModeMediaQuery = JSON.parse(document.querySelector('#darkModeMediaQuery').innerText);
@@ -253,7 +253,9 @@ onload = () => {
         // Always show exp & out.
         document.querySelector('#exp div').innerText = data.Exp;
         document.querySelector('#out div').innerText = data.Out;
-        updateDiff(data.Exp, data.Out, data.Argv);
+
+        document.querySelector("#diff-content").innerHTML = diffHTML(hole, data.Exp, data.Out, data.Argv);
+        diff.style.display = data.Exp === data.Out ? 'none' : 'block'
 
         status.className = data.Pass ? 'green' : 'red';
         status.style.display = 'block';
@@ -387,174 +389,4 @@ async function refreshScores() {
 
 function getScoring(str, index) {
     return scorings[index] == 'Bytes' ? new TextEncoder().encode(str).length : strlen(str);
-}
-
-function getLineChanges(before, after) {
-    const includeArgs = getDiffType() == 'arg'
-    if (includeArgs) {
-        const out = []
-        const splitBefore = lines(before)
-        const splitAfter = lines(after)
-        const compareOpts = {
-            sensitivity: shouldIgnoreCase() ? 'accent' : 'base'
-        }
-        for (let i=0; i<Math.max(splitBefore.length, splitAfter.length); i++) {
-            const a = splitBefore[i]
-            const b = splitAfter[i]
-            // https://stackoverflow.com/a/2140723/7481517
-            const linesEqual = 0 === a.localeCompare(b, undefined, compareOpts);
-            if (linesEqual) {
-                out.push({
-                    count: 1,
-                    value: a + '\n'
-                })
-            } else {
-                for (let [k,v] of [['removed', a], ['added', b]]) {
-                    if (v !== undefined) {
-                        const prev = out[out.length - 1]
-                        if (prev && prev[k]) {
-                            prev.count++;
-                            prev.value += v + '\n'
-                        } else {
-                            out.push({
-                                count: 1,
-                                [k]: true,
-                                value: v + '\n'
-                            })
-                        }
-                    }
-                }
-            }
-        }
-        return out
-    } else {
-        return Diff.diffLines(before, after, {
-            ignoreCase: shouldIgnoreCase()
-        })
-    }
-}
-
-function updateDiff(exp, out, argv) {
-    diff.style.display = exp === out ? 'none' : 'block'
-    let html = `<div class='diff-title-bg'></div>
-        <div class='diff-title-bg'></div>
-        <h3 class='diff-output'>Output</h3>
-        <div class='diff-title-bg'></div>
-        <h3 class='diff-expected'>Expected</h3>`;
-    let pos = {
-        left: 1,
-        right: 1
-    };
-    const changes = getLineChanges(out, exp);
-    let pendingChange = null;
-    for (let i = 0; i < changes.length; i++) {
-        const change = changes[i]
-        pos.isLastDiff = i === changes.length - 1
-        if (change.added || change.removed) {
-            if (pendingChange === null) {
-                pendingChange = change;
-            } else {
-                html += getDiffRow(pendingChange, change, pos, argv);
-                pendingChange = null;
-            }
-        } else {
-            if (pendingChange) {
-                html += getDiffRow(pendingChange, {}, pos, argv);
-                pendingChange = null;
-            }
-            html += getDiffLines(change, change, pos, argv);
-        }
-    }
-    if (pendingChange) {
-        html += getDiffRow(pendingChange, {}, pos, argv);
-    }
-    diff.querySelector("div").innerHTML = html;
-}
-
-function getDiffRow(change1, change2, pos, argv) {
-    change2.value ??= ''
-    change2.count ??= 0
-    const left = change1.removed ? change1 : change2
-    const right = change1.added ? change1 : change2
-    return getDiffLines(left, right, pos, argv)
-}
-
-function getDiffLines(left, right, pos, argv) {
-    const leftSplit = lines(left.value);
-    const rightSplit = lines(right.value);
-    if (!(pos.isLastDiff && hole === "quine")) {
-        // ignore trailing newline
-        if (leftSplit[leftSplit.length - 1] === '') leftSplit.pop();
-        if (rightSplit[rightSplit.length - 1] === '') rightSplit.pop();
-    }
-    let s = ''
-    const diffOpts = {
-        ignoreCase: shouldIgnoreCase()
-    }
-    for (let i=0; i<Math.max(leftSplit.length, rightSplit.length); i++) {
-        const leftLine = leftSplit[i];
-        const rightLine = rightSplit[i];
-        const charDiff = Diff.diffChars(leftLine ?? '', rightLine ?? '', diffOpts);
-        // subtract 1 because the lines start counting at 1 instead of 0
-        const arg = argv[i + pos.right - 1]
-        if (arg !== undefined) {
-            s += `<div class='diff-arg'>${arg}</div>`
-        }
-        if (leftLine !== undefined) {
-            s += `<div class='diff-left-num'>${i + pos.left}</div>
-                <div class='diff-left${left.removed?' diff-removal':''}'>${renderCharDiff(charDiff, false)}</div>`
-        }
-        if (rightLine !== undefined) {
-            s += `<div class='diff-right-num'>${i + pos.right}</div>
-                <div class='diff-right${right.added?' diff-addition':''}'>${renderCharDiff(charDiff, true)}</div>`
-        }
-    }
-    pos.left += left.count;
-    pos.right += right.count;
-    return s
-}
-
-function renderCharDiff(charDiff, isRight) {
-    let html = ''
-    for (let change of charDiff) {
-        if (change.added && isRight) {
-            html += `<span class='diff-char-addition'>${change.value}</span>`
-        } else if (change.removed && !isRight) {
-            html += `<span class='diff-char-removal'>${change.value}</span>`
-        } else if (!change.added && !change.removed) {
-            html += change.value;
-        }
-    }
-    return html
-}
-
-function getDiffType() {
-    switch (hole) {
-        case 'arabic-to-roman':
-        case 'arrows':
-        case 'css-colors':
-        case 'emojify':
-        case 'fractions':
-        case 'intersection':
-        case 'levenshtein-distance':
-        case 'musical-chords':
-        case 'ordinal-numbers':
-        case 'roman-to-arabic':
-        case 'spelling-numbers':
-        case 'united-states':
-            // { | category = Transform }
-            // - {Pangram Grep, QR Decoder, Seven Segment, Morse Decoder, Morse Encoder}
-            // + {Fractions, Intersection}
-            return 'arg'
-        default:
-            return 'line'
-    }
-}
-
-function shouldIgnoreCase() {
-    return hole === "css-colors"
-}
-
-function lines(s) {
-    return s.split(/\r\n|\n/)
 }
