+++ views/hole-ng.html
@@ -151,7 +151,7 @@
         <div id=arguments></div>
         <div id=output   ></div>
         <div id=expected ></div>
-        <div id=diff     ></div>
+        <div id=diff     ><div id=diff-content></div></div>
         <div id=errors   ></div>
     </div>
 </main>
