+++ views/hole.html
@@ -131,7 +131,7 @@
         <aside id=arg><h3 class='aside-title'>Arguments</h3><div></div></aside>
         <aside id=exp><h3 class='aside-title'>Expected</h3><div></div></aside>
         <aside id=out><h3 class='aside-title'>Output</h3><div></div></aside>
-        <aside id=diff><h3 class='aside-title'>Diff</h3><div></div></aside>
+        <aside id=diff><h3 class='aside-title'>Diff</h3><div id=diff-content></div></aside>
     </div>
 </main>
